Update Bundle config directory
Currently, we use `/bundle_app_config` as the Bundle config location.
This works for fine for `root` user, but for non-root users, it won't be
able create the directory as the user doesn't have enough permission to
perform the operation.

This commit updates the Bundle config directory from
`/bundle_app_config` to `/tmp/bundle_app_config` which going to fix the
permission issue.